SWAR-39
Changed some lines in the h file to make it consistent with the develop brach which was causing the build to fail

+++ src/behaviours/src/Utilities.cpp
@@ -25,5 +25,5 @@ int Utilities::saturation_check(int direction, int sat) {
 
 
 float Utilities::infinity_norm(Point input) {
-	return fmax(std::abs(input.x), std::abs(input.y));
+	return fmaxf(std::abs(input.x), std::abs(input.y));
 }
